Close connection only in transaction-less case
An existing transaction may hold an open transaction without a transaction can't exist.
If the connection is closed, the transaction is A) broken/rolled back and B) doesn't know that the transaction was closed. Any other use of the transaction would result in an exception.

Solution: only close the transaction if no transaction was available.